Pull directly from remote without extra latency
